third_party/WebKit/LayoutTests/screen_orientation/lockOrientation-bad-argument-expected.txt

Validates that screen.lockOrientation() returns false on bad input

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS screen.orientation is "portrait-primary"
PASS screen.lockOrientation('invalid-orientation') is false
PASS screen.orientation is "portrait-primary"
PASS screen.lockOrientation(null) is false
PASS screen.orientation is "portrait-primary"
PASS screen.lockOrientation(undefined) is false
PASS screen.orientation is "portrait-primary"
PASS screen.lockOrientation(123) is false
PASS screen.orientation is "portrait-primary"
PASS screen.lockOrientation(window) is false
PASS screen.orientation is "portrait-primary"
PASS screen.lockOrientation() threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': 1 argument required, but only 0 present..
PASS screen.orientation is "portrait-primary"
PASS successfullyParsed is true

TEST COMPLETE
